Updated with new pattern to remove the data source
