+++ core/database/src/main/java/org/mozilla/social/core/database/DatabaseModule.kt
@@ -3,8 +3,6 @@ package org.mozilla.social.core.database
 import androidx.room.Room
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
-import org.mozilla.social.core.database.dao.PollsDao
-import org.mozilla.social.core.database.datasource.PollLocalDataSource
 
 val databaseModule = module {
     single {
@@ -14,5 +12,4 @@ val databaseModule = module {
         ).build()
     }
     single { get<SocialDatabase>().pollDao() }
-    single { PollLocalDataSource(get()) }
 }
