+++ core/repository/mastodon/src/main/java/org/mozilla/social/core/repository/mastodon/StatusRepository.kt
@@ -1,20 +1,17 @@
 package org.mozilla.social.core.repository.mastodon
 
-import androidx.room.withTransaction
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import org.mozilla.social.core.database.SocialDatabase
-import org.mozilla.social.core.database.datasource.PollLocalDataSource
-import org.mozilla.social.core.network.mastodon.StatusApi
-import org.mozilla.social.core.model.PollVote
-import org.mozilla.social.core.repository.mastodon.model.context.toExternalModel
-import org.mozilla.social.core.repository.mastodon.model.status.toDatabaseModel
-import org.mozilla.social.core.repository.mastodon.model.status.toExternalModel
-import org.mozilla.social.core.repository.mastodon.model.status.toNetworkModel
 import org.mozilla.social.core.model.Context
 import org.mozilla.social.core.model.Poll
+import org.mozilla.social.core.model.PollVote
 import org.mozilla.social.core.model.Status
 import org.mozilla.social.core.model.request.StatusCreate
+import org.mozilla.social.core.network.mastodon.StatusApi
+import org.mozilla.social.core.repository.mastodon.model.context.toExternalModel
+import org.mozilla.social.core.repository.mastodon.model.status.toExternalModel
+import org.mozilla.social.core.repository.mastodon.model.status.toNetworkModel
 
 class StatusRepository(
     private val statusApi: StatusApi,
